Associate Movement With Categorie Fix #25

+++ PTZ.HomeManagement.MyFinance.Core/Models/CategoryBankAccountMovement.cs
@@ -6,9 +6,9 @@ namespace PTZ.HomeManagement.MyFinance.Models
 {
     public class CategoryBankAccountMovement
     {
-        public int BankAccountMovementId { get; set; }
+        public long BankAccountMovementId { get; set; }
         public BankAccountMovement BankAccountMovement { get; set; }
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public Category Category { get; set; }
     }
 }
